IPv6 support
After HTTP::Daemon gains IPv6 support, RPC-XML tests will fail because
the RPC server will listen on IPv6 socket and clients will connect
from an IPv6 address on systems where IPv6 is available.

This patch changes tests to handle IPv6 addresses correctly.

This patch also adds "peerfamily" key to RPC::XML::Server connection
objects because a packed network address in "peeraddr" key is
umbiguous and cannot be unpacked without previous knowledge of the
address family that it encodes.

There are still some hard coded AF_INET calls in test to find used or
unused ports, but the approach is defective by design because of
a possible race between the check for a port and subsequent use of the
port. RPC::XML::Server would have to learn how to run on a user-supplied
socket first to fix the races correctly.

<https://rt.cpan.org/Public/Bug/Display.html?id=120472>

Signed-off-by: Petr Písař <ppisar@redhat.com>